🚚 Move: CustomClearXmarkTextField 파일 분리
- Login / PhoneNumber 에 Components 그룹 생성
- 해당 그룹에 파일 생성
